some work but not finish

+++ libft/ft_atoi.c
@@ -16,9 +16,9 @@
 // fonction qui transforme une string de char en int,
 // detecte les spaces, marche avec les nbrs negatifs.
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	int				res;
+	long				res;
 	int				sign;
 	unsigned int	i;
